PythonServer/requirements.txt
fastapi
uvicorn
pinecone-client
pdfplumber
langchain
langchain-pinecone
langchain-aws
pydantic
python-dotenv
langchain_text_splitters
psycopg[binary]
asyncpg
langsmith
langchain_core
langgraph
langgraph-prebuilt
langgraph-checkpoint-postgres
tiktoken
python-multipart
pymupdf